feat: Currency list API
Make API invoker.
Create currency list API wrapper function.
